requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
Pillow==9.0.0
torch==1.11.0+cpu
torchvision==0.12.0+cpu
gunicorn==20.1.0
facenet-pytorch==2.5.2
flask
